+++ config/settings.py
@@ -1,11 +1,11 @@
 """
 /config/settings.py
-Server Monitoring System v8.62.65
+Server Monitoring System v8.62.66
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Application settings - default values
 Система мониторинга серверов
-Версия: 8.62.65
+Версия: 8.62.66
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Настройки приложения - значения по умолчанию
@@ -22,9 +22,9 @@ from lib.utils import is_proxmox_server
 DEBUG_MODE = False
 
 # Версия приложения
-APP_VERSION = "8.62.65"
+APP_VERSION = "8.62.66"
 ANDROID_MIN_SUPPORTED_VERSION = "8.59.10"
-ANDROID_LATEST_VERSION = "8.62.65"
+ANDROID_LATEST_VERSION = "8.62.66"
 
 
 # Matrix-уведомления (доп. канал к Telegram)