+++ extensions/extension_manager.py
@@ -1,11 +1,11 @@
 """
 /extensions/extension_manager.py
-Server Monitoring System v8.62.65
+Server Monitoring System v8.62.66
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Extension Manager for Monitoring
 Система мониторинга серверов
-Версия: 8.62.65
+Версия: 8.62.66
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Менеджер расширений для мониторинга