+++ modules/improved_mail_monitor.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 """
 /modules/improved_mail_monitor.py
-Server Monitoring System v8.62.65
+Server Monitoring System v8.62.66
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Mailbox monitoring
 Система мониторинга серверов
-Версия: 8.62.65
+Версия: 8.62.66
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Мониторинг почтового ящика